backend/requirements-dev.txt
-r requirements.txt
pytest==8.2.2
pytest-flask==1.3.0
factory-boy==3.3.0
freezegun==1.5.1
coverage==7.5.3
pytest-cov==5.0.0
